user can now specify a bitmask for activate_line and clear_line
activate_line and clear_line now accept 'lines' and 'bitmask' as keyword arguments to the function.  The user must specify one, but not both.
If lines is specified, the user can enter a list of lines whose values are between 1 and 8.  If bitmask is specified, the user can specify a value between 0 and 255.